requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
apipkg==1.4               # via execnet
contextlib2==0.5.1        # via raven
coverage==4.0.3
decorator==4.0.9          # via ipython, traitlets
dj-database-url==0.4.0
django-getenv==1.3.1
django-redis==4.3.0
django==1.9.2
execnet==1.4.1            # via pytest-xdist
flake8==2.5.4
ipdb==0.9.0
ipython-genutils==0.1.0   # via traitlets
ipython==4.1.1            # via ipdb
mccabe==0.4.0             # via flake8
newrelic==2.60.0.46
path.py==8.1.2            # via pickleshare
pep8==1.7.0               # via flake8
pexpect==4.0.1            # via ipython
pickleshare==0.6          # via ipython
psycopg2==2.6.1
ptyprocess==0.5.1         # via pexpect
py==1.4.31                # via pytest, pytest-xdist
pyflakes==1.0.0           # via flake8
pytest-cov==2.2.1
pytest-django==2.9.1
pytest-html==1.8.0        # via pytest-selenium
pytest-selenium==1.1
pytest-variables==1.3     # via pytest-selenium
pytest-xdist==1.14
pytest==2.8.7
python-logstash==0.4.6
pytz==2015.7
raven==5.10.2
redis==2.10.5             # via django-redis
requests==2.9.1           # via pytest-selenium
selenium==2.52.0          # via pytest-selenium
simplegeneric==0.8.1      # via ipython
traitlets==4.1.0          # via ipython
uwsgi==2.0.12

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via ipdb, ipython
